Pass Data Between Two Activities in Android application Using Intent
An Intent is an object that provides runtime binding between separate components, such as two activities. The Intent represents an app’s intent to do something. You can use intents for a wide variety of tasks, but in this lesson, your intent starts another activity.

The easiest way to do this would be to pass message to the DisplayActivity in the Intent use to start the activity:

        Intent intent = new Intent(this, DisplayActivity.class);
        intent.putExtra("name_info", message);
        startActivity(intent);

Access that intent on DisplayActivity:

        Bundle get_name_from_main = getIntent().getExtras();
        if (get_name_from_main != null) {
            name = get_name_from_main.getString("name_info");
            //The key argument here must match that used in the other activity
        }

Hope It'll help you.